requirements.txt
spotipy>=2.23
yt-dlp>=2024.1
mutagen>=1.47
typer>=0.12
rich>=13.0
python-dotenv>=1.0
requests>=2.31
